Use XDG env vars for user data/confg/cache locations
We used glib.get_user_data_dir in the python part and XDG_DATA_HOME in
the vim part, which *might* introduce an asymmetry. For simplicitly and
transparency, we use $XDG_* throughout.

NOTE: This might cause kzrnote to not find old notes, they need to be
moved to the correct dir manually.